requirements.txt
chromadb
Flask
Flask-Cors
"Flask-Limiter[mongodb]"
llama-index-core
llama-index-readers-file
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-storage-index-store-mongodb
llama-index-vector-stores-chroma
motor
pymongo
python-dotenv
